Refactor the custom script file structure
