+++ components/custom-script.tsx
@@ -3,7 +3,7 @@
  */
 
 import { ScriptHTMLAttributes, useEffect } from "react";
-import { requestIdleCallback } from "./request-idle-callback";
+import { requestIdleCallback } from "../lib/request-idle-callback";
 
 const loadScript = (props: ScriptHTMLAttributes<HTMLScriptElement>): void => {
   const { dangerouslySetInnerHTML } = props;
